# test/crimson/seastore/test_omap_manager: add unit test for clear