@chanyoung 2025-07-29
jenkins test make check arm64