@athanatos 2025-08-07
https://pulpito.ceph.com/sjust-2025-08-05_23:22:50-crimson-rados-wip-sjust-seastore-testing-2025-08-04-1754319725-distro-crimson-debug-gibba/

failures seem unrelated